third_party/blink/web_tests/external/wpt/svg/geometry/svg-get-bounding-client-rect-in-non-rendered-elements-expected.txt

This is a testharness.js-based test.
[FAIL] Get Bounding Client Rect
  assert_equals: g1 expected 10 but got 45
Harness: the test ran to completion.
